+++ src/collective/cover/tiles/edit_widgets/interfaces.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from z3c.form import interfaces
+
+
+class IMoreLinkWidget(interfaces.IWidget):
+    """More... link widget."""